feat: started to split file logic into a mixin
